@@ -5,21 +5,21 @@
   "homepage": "https://angelzbg.github.io/portfolio",
   "dependencies": {
     "@types/jest": "^26.0.20",
-    "@types/node": "^14.14.22",
-    "@types/react": "^17.0.0",
+    "@types/node": "^14.14.25",
+    "@types/react": "^17.0.1",
     "@types/react-dom": "^17.0.0",
     "@types/react-router": "^5.1.11",
     "@types/react-router-dom": "^5.1.7",
-    "bootstrap": "^4.6.0",
-    "jquery": "^3.4.1",
-    "mobx": "^6.0.4",
-    "mobx-react": "^7.0.5",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
+    "bootstrap": "4.6.0",
+    "jquery": "^3.5.1",
+    "mobx": "^6.1.5",
+    "mobx-react": "^7.1.0",
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1",
     "react-feather": "^2.0.9",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "3.1.2"
+    "react-scripts": "^4.0.2"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -45,8 +45,8 @@
     ]
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^4.14.0",
-    "@typescript-eslint/parser": "^4.14.0",
+    "@typescript-eslint/eslint-plugin": "^4.14.2",
+    "@typescript-eslint/parser": "^4.14.2",
     "gh-pages": "^2.2.0",
     "typescript": "^4.1.3"
   }
