@@ -1,15 +1,15 @@
-const icon_nodejs = require('../../images/icon_nodejs.png');
-const icon_javascript = require('../../images/icon_javascript.png');
-const icon_css = require('../../images/icon_css.png');
-const icon_html = require('../../images/icon_html.png');
-const icon_github = require('../../images/icon_github.png');
-const icon_firebase = require('../../images/icon_firebase.png');
-const icon_express = require('../../images/icon_express.png');
-const icon_mongodb = require('../../images/icon_mongodb.png');
-const icon_reactjs = require('../../images/icon_react.png');
-const cert_mini_softuni_basics = require('../../images/cert_mini_softuni_basics.png');
-const cert_mini_softuni_fundamentals = require('../../images/cert_mini_softuni_fundamentals.png');
-const cert_mini_softuni_advanced = require('../../images/cert_mini_softuni_advanced.png');
+import icon_nodejs from '../../images/icon_nodejs.png';
+import icon_javascript from '../../images/icon_javascript.png';
+import icon_css from '../../images/icon_css.png';
+import icon_html from '../../images/icon_html.png';
+import icon_github from '../../images/icon_github.png';
+import icon_firebase from '../../images/icon_firebase.png';
+import icon_express from '../../images/icon_express.png';
+import icon_mongodb from '../../images/icon_mongodb.png';
+import icon_reactjs from '../../images/icon_react.png';
+import cert_mini_softuni_basics from '../../images/cert_mini_softuni_basics.png';
+import cert_mini_softuni_fundamentals from '../../images/cert_mini_softuni_fundamentals.png';
+import cert_mini_softuni_advanced from '../../images/cert_mini_softuni_advanced.png';
 
 export interface ISoftuniIcon {
   icon: any;
