@@ -2,7 +2,7 @@ import React from 'react';
 import './styles/articles.css';
 import RedditArticle from './components/RedditArticle';
 
-export default ({ isOpen }: { isOpen: boolean }) => (
+const Articles = ({ isOpen }: { isOpen: boolean }) => (
   <div className="dropdown-menu-wrapper articles-dropdown-wrapper" style={{ display: isOpen ? 'block' : 'none' }}>
     <div className="dropdown-menu-wrap articles-dropdown-menu-wrap">
       <div className="articles scrollbar">
@@ -11,3 +11,5 @@ export default ({ isOpen }: { isOpen: boolean }) => (
     </div>
   </div>
 );
+
+export default Articles;
